added new directories by categories and moved existing tests into their respective directories
