+++ src/Array/slice.js
